Fixed bug that could cause wrong data for some initial values of Aether
The error was in the validation of the initial value. It only checked
whether the maximum value that would result from using the passed
initial value would exceed the range of the primitive used. But it
didn't take into account that during the calculation of the algorithm a
greater value would show up and overflow the primitive type.